use an Intent to switch to another Activity
This shows how we can set up an Intent to switch over to the
RecipeActivity. Note how we pass along the recipe title in the
Intent, so that the receiving Activity will know what to use
for a title.

Note how we also need to go adding an intent-filter to the
AndroidManifest.xml file.