implement second test
Adding a simple special case to a method to make yet another
test pass is called "sliming". It is encouraged. The intent
is for more ambitious improvements to happen in the "refactor"
phase of the TDD cycle. In trying to make the test pass,
you shouldn't have to think about the big picture.