dependency injection step 3
Extract an interface from the concrete class.

This is the final step. Now it doesn't matter if we
pass in our original concrete implementation, or some
other concrete implementation. All they have to do is
implement the interface.

This is a deep principle. "Program towards an interface"
gets us very far. It's a sensible use of OOP.

There's a refactoring tool called "Extract interface"
that does this.