write tests using a mock class
Now that we've extracted the interface, we can write our
own concrete class that conspires to make testing easy.
A concrete class that functions as a kind of testing spy
is called a "mock class".

Note that since we have complete control over this new
class, we can give it arbitrary methods to simulate
various changes in our mocked external environment. The
method moveNorth() is an example of this. We could just
as easily simulate the passage of time, for example, by
mocking up a time class.